[MCCS-331] Fix dsconfig workdir and remove json2tango wrapper
Many helm charts use relative paths for dsconfig data files,
so ensure the container starts in `/` instead of `/app`,
now that dsconfig is based on ska-python-runtime.

Also remove the `json2tango` wrapper script - it was not correcty
added to the PATH previously (tango-dsconfig:1.5.0), so was not
actually used.  No need for it now either.  The Helm chart
`bootstrap.sh` script already checks the return code of 2.