WOM-229 tests: Remove reference to CmdArgType.DevInt
DevInt was removed from the cppTango CmdArgType enum as it is unused in
the kernel, this has been reflected in pytango as well.  It should not
have been possible to use DevInt in tango 9.4.x, so this should not
break real code.

Cherry-picked from ca4a9c8 (WOM-134: Remove reference to
CmdArgType.DevInt, 2023-09-26).