WOM-229 Makefile: Use commit tags for images in the CI
Before this commit we referenced images via their release tag, however,
when we go to test the images with the `k8s-test` CI for a second time
(i.e. because we have made some changes to the image) Kubernetes will not
not pull the latest version of the image, but instead use the local
version it finds with the release tag which was build during the
original run of the CI.

With this change, we ensure that the CI is always testing the images
generated by this CI pipeline.