WOM-243: Use --check-device-exported for readinessProbe
We need to use --check-device-exported over --ping-device because
--ping-device uses the ORBendPointPublish address and the container is
only available for the published hostname _after_ the container has been
determined ready.

To ensure that the device is exported is caused by _this_ pod starting
by first unexporting all devices before we launch the server.

We also ensure that the entire command is always passed as a single
argument to `sh -c` as sh -c only uses the first argument.