WOM-459: Rework pytango-runtime and remove pytango-builder
pytango-builder is no longer needed as pytango can be installed from a
wheel and a python build images are provided by ska-base-image.

For the pytango-runtime image, we include the contents of the
tango-admin image so that tango_admin is available in images containing
Tango device servers.

We no longer need pip.conf because this should be handled by the
ska-base-images.