WOM-459: Add Makefile rules for building locally
These rules use the dependency graph defined by the Dockerfiles so that
images are built in the correct order.

We also use a local tag by default so that the tags do not change as
this are committed or the repository becomes "dirty".

This makes it much easier to work on the images locally.