WOM-515: Update image version numbers
As we are not updating any upstream packages with the 0.5.0 release, we
in general, just need to bump the patch version.  In some cases, I have
updated the major/minor versions as well to better match the upstream
application we are installing.  This images had been previously
mislabeled.

For example, dsconfig (version number not updated here) was previously
tagged as 1.5.13, however, we were installing version 1.7.1.  For this
release of ska-tango-images we are calling the image 1.7.1.

Similarly, for the TangoRestServer we were installing version 1.22.0 but
were labeling the image as 1.14.9.