WOM-517: Add rest-server and deprecate tango-rest
The tango-rest image is using supervisord which I think is the wrong
approach for a device server image.  We should let docker compose manage
the lifecycle of the service by providing an appropriate healthcheck.

As removing supervisord is a breaking change, I have decided to rename
the image to rest-server.  As the previous release of tango-rest was
broken, I feel it is best to release a fixed version before we remove
it in the next release.