WOM-517: Move substitution generation to ska_tango_images python package
This anticipates us using this logic to generate files from an
additional script.  The version information logic could probably also be
used in the tests, however, this has been left to do later.

As the documentation build now requires the ska_tango_images package, we
have to adjust a few things to make sure it is installed.