WOM-567: Update to cppTango 10.0.0 with opentelemetry
Adding opentelemetry support involves adding a lot more dependencies, it
is much easier to manage this in the same Dockerfile rather than having
the separate ska-tango-image-tango-dependencies image.

The ska-tango-image-tango-dependencies image was just an implementation
detail and was never advertised in the catalogue.  It is also of
questionable utility to anyone outside of this project since it is
specifically for building cppTango. Therefore, it is safe to remove
this.

To keep track of the system runtime dependencies which have to be
installed this commit adds a /runtime_deps.txt to the image.  This means
the required packages only have to be listed in one place.

We no longer include any tools in the ska-tango-images-tango-cpp image,
so downstream images need to install their own.  This requires adding
pkg-config to ska-tango-images-tango-test and
ska-tango-images-tango-databaseds.

Also, libhdbpp does not build against cppTango 10.0.0, but we can work
around this by removing the FindTango.cmake file.